Support remote hosts and add a force_update option
Prior to this, the script was unable to work with remote hosts because
of the check that determines whether Kodi was running or not checked the
local process table and would fail if not found.

Conditional logic added so that, if running locally, we'll check that
kodi is running before trying to hit its API. If not running locally,
we'll always try to his its API.

Also added the force_update option that , if 'yes', will always try to
his the API regardless of if Kodi is local or not.